requirements.txt
streamlit
google-generativeai
PyPDF2
python-dotenv
plotly
matplotlib
pandas  # Optionnel si ds_course dépend de pandas
